move hasBuildToolConflicts to utils.ts
